Cleaning

+++ Moduuli03/kuha.py
@@ -1,5 +1,4 @@
-#37cm
-kuha_length = int(input("Syötä kuhan pituus senttimetreinä: "))
+kuha_length = float(input("Syötä kuhan pituus senttimetreinä: "))
 alin_mitta = 37
 if kuha_length < alin_mitta:
     print(f"Kuha on alamittainen.\nLaske kuha takaisin järveen.")
